Added BaseSplitter.php abstract class for all slitters Added Gaufrette in clients in BackupCommand and fixed gaufrette upload case
